llm-v2/requirements.txt
matplotlib
pytest
spacy
tika
torch
torchvision
pyspark
fastapi
setuptools
PyMySQL
numpy
starlette
ray
ray[serve]
requests
pykwalify
PyYAML
faiss-cpu
sentence-transformers
elasticsearch
pandas
PyArrow
pydantic
typing

